split up src/data.ts
